Do biquad filtering on floats instead of doubles.
There is a slight difference in the output samples with this change,
but on GBA it's usually only ~37% of the samples are different, and in
most cases it's within +/-1 of the result computed with doubles.
The largest difference I saw was +/-4, but that was infrequent.

In any case, I can't hear any difference.

+++ src/gb/audio/Audio.h
@@ -148,7 +148,7 @@ private:
 
     // Q values are for a 4th order cascaded Butterworth lowpass filter.
     // Obtained from http://www.earlevel.com/main/2016/09/29/cascading-filters/.
-    static constexpr std::array<double, 2> q{0.54119610, 1.3065630};
+    static constexpr std::array<float, 2> q{0.54119610f, 1.3065630f};
     std::vector<Common::Biquad> biquads;
 
     void UpdatePowerOnState();
